requirements.txt
asgiref==3.2.2
asn1crypto==0.24.0
attrs==19.3.0
autobahn==19.9.3
Automat==0.7.0
backports.functools-lru-cache==1.6.1
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.7.0.1
bcrypt==3.1.7
beautifulsoup4==4.8.1
boto==2.0
boto3==1.10.35
botocore==1.13.35
cached-property==1.5.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cliapp==1.20170827
cmdtest==0.32
CodeConvert==3.0.2
configparser==4.0.2
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cssutils==1.0.2
Django==1.11.11
django-admin==2.0.0
django-allauth==0.40.0
django-bootstrap4==1.0.1
django-common-helpers==0.9.2
django-crispy-forms==1.8.1
django-cron==0.5.1
django-crontab==0.7.1
django-elasticsearch-dsl==7.1.0
django-elasticsearch-dsl-drf==0.20.3
django-excel-response2==2.0.8
django-filter==2.0.0
django-inlinecss==0.3.0
django-models-ext==1.1.9
django-nine==0.2.2
django-rest-auth==0.9.5
django-rest-swagger==2.0.7
django-restframework==0.0.1
django-short-url==1.1.4
django-six==1.0.4
django-storages==1.8
django-widget-tweaks==1.4.5
djangorestframework==3.9.4
docker==4.1.0
docker-compose==1.25.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
enum34==1.1.6
functools32==3.2.3.post2
furl==2.1.0
future==0.18.2
futures==3.3.0
get==2019.4.13
gyp==0.1
httpie==1.0.3
idna==2.8
importlib-metadata==1.4.0
iniparse==0.4
ipaddress==1.0.23
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==3.2.0
keyring==10.6.0
keyrings.alt==3.0
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==5.0.0
oauthlib==3.1.0
openapi-codec==1.3.2
orderedmultidict==1.0.1
paramiko==2.7.1
path.py==11.5.2
pathlib==1.0.1
pathlib2==2.3.5
Pillow==6.2.1
post==2019.4.13
postman==0.6.0
psycopg2==2.8.4
public==2019.4.13
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.1
Pygments==2.4.2
pygobject==3.26.1
PyJWT==1.4.2
pyliblzma==0.5.3
PyNaCl==1.3.0
pynliner==0.8.0
pyrsistent==0.15.7
PySide==1.2.2
pysqlite==1.0.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-openid==2.2.5
pytz==2019.3
pyxdg==0.25
PyYAML==5.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
rpm==4.14.1
s3transfer==0.2.1
scandir==1.10.0
screen==1.0.1
SecretStorage==2.3.1
shell==1.0.1
shortuuid==0.5.0
simplejson==3.17.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.2.4
subprocess32==3.5.4
texttable==1.6.2
TimeConvert==1.5.0
ttystatus==0.38
txaio==18.8.1
typing==3.7.4.1
tzlocal==2.0.0
uritemplate==3.0.0
urlgrabber==3.10.2
urllib3==1.25.8
virtualenv==16.7.8
websocket-client==0.57.0
xlwt==1.3.0
yum-metadata-parser==1.1.4
zipp==1.0.0
